+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: (TON) Replace Orbs servers with dRPC
+
 ## 4.77.2 (2026-03-24)
 
 - changed: Update chain-registry package